Better handling of `CircuitData` instead of working with Python QuantumCircuit (buggy)
