+++ crates/qpy/src/lib.rs
@@ -36,7 +36,6 @@ mod value;
 /// be re-exposed directly to public Python space.
 pub fn qpy(module: &Bound<PyModule>) -> PyResult<()> {
     module.add_function(wrap_pyfunction!(circuit_writer::py_write_circuit, module)?)?;
-    module.add_function(wrap_pyfunction!(circuit_reader::py_read_circuit, module)?)?;
     module.add_function(wrap_pyfunction!(interface::py_dump_qpy, module)?)?;
     module.add_function(wrap_pyfunction!(interface::py_load_qpy, module)?)?;
     Ok(())
